Moved risks related functions from fleetService to risksService
